refactor: product 엔티티 위치 이동
