+++ src/main/java/gift/domain/product/ProductController.java
@@ -3,7 +3,6 @@ package gift.domain.product;
 import gift.global.response.ResponseMaker;
 import gift.global.response.ResultResponseDto;
 import gift.global.response.SimpleResultResponseDto;
-import gift.entity.Product;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
