+++ src/test/java/gift/ProductTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.domain.product.ProductDTO;
-import gift.entity.Product;
+import gift.domain.product.Product;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
